% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonms.R
\name{anonms}
\alias{anonms}
\alias{templ_anonms}
\title{Steve's Anonymous Manuscript Template}
\usage{
anonms(...)

templ_anonms()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
The R Markdown template Steve used once for an anonymous manuscript submission that needed to look an exact way. I've used this only once
for a manuscript submission to a particular journal. It was not published. I have not had the occasion to submit there again.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template. Honestly, this looks like a boilerplate R Markdown template. I think \code{runhead} is the only parameter
unique to this document. It's also mandatory. Other fields that you can see in the skeleton (e.g. \code{title}, \code{name}, \code{affiliation})
are borrowed from other templates and explicated there.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{runhead} \tab A shortened version of \code{title} to treat as a running head \cr
}
}

