% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{article}
\alias{article}
\title{Steve's First Article/Manuscript Template}
\usage{
article(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for academic articles/manuscripts. For more information, see here:
\url{http://svmiller.com/blog/2020/09/another-rmarkdown-article-template/}.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{name} \tab name of the author \cr
   \code{affiliation} \tab the affiliation of the author \cr
   \code{thanks} \tab A place to thank people for their input, identify corresponding author, etc. \cr
   \code{abstract} \tab include an abstract of the paper here \cr
   \code{keywords} \tab some keywords for the manuscript \cr
   \code{anonymous} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, scrubs all identifying author info for peer review \cr
   \code{removetitleabstract} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, removes title/abstract for peer review (if you'd like) \cr
   \code{appendix} \tab optional, but specifies appendix figure/table prefices \cr
   \code{endnotes} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, makes R Markdown footnotes as endnotes \cr
   \code{twocol} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, makes document two-column format \cr
}
}

