% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{beamer}
\alias{beamer}
\alias{templ_beamer}
\title{Steve's Beamer Template}
\usage{
beamer(...)

templ_beamer()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:beamer_presentation]{rmarkdown::beamer_presentation}}.}
}
\description{
This is the R Markdown template I use for in-class or in-person presentations. For
a discussion of some of its features, see here:
\url{http://svmiller.com/blog/2019/08/r-markdown-template-beamer-presentations/}.
}
\details{
Note that \code{xelatex} is not required for this template but is heavily suggested. If
you elect to use \code{xelatex}, you \emph{must} supply something in \verb{titlefont:} and
\verb{mainfont:} that your particular LaTeX installation will recognize. See discussion
here: \url{https://github.com/svmiller/stevetemplates/issues/4}
}
\section{About YAML header fields}{
\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab Title of the presentation \cr
   \code{subtitle} \tab Subtitle of the presentation. Optional, but useful. For in-class lectures, I put the class name here. \cr
   \code{institute} \tab List your department/college here. Optional. \cr
   \code{titlegraphic} \tab Put a path to university watermark here. Optional. \cr
   \code{scaletg} \tab Scale \code{titlegraphic} relative to the width of the document. Optional, defaults to .3 \cr
   \code{titlegraphshift} \tab Optional LaTeX commands to move around \code{titlegraphic}. Manually escape any LaTeX code you put here. Use with care. \cr
   \code{make149} \tab Logical, gives 14:9 dimension to presentation (my favorite). Optional, defaults to FALSE. \cr
   \code{mainfont} \tab The main font to use with \code{xelatex}. I use \verb{Open Sans} here. Optional, but recommended. \cr
   \code{titlefont} \tab The font for slide titles. I use \verb{Titillium Web} here. Optional, but recommended. \cr
   \code{primarycolor} \tab Specifies a "primary" color in six-digit RGB color. Defaults to \verb{002F5F}. Optional. \cr
   \code{secondarycolor} \tab Specifies a "secondary" color in six-digit RGB color. Defaults to \verb{9BB2CE}. Optional. \cr
   \code{no_toc_after} \tab When \code{TRUE}, disables table of contents at end of document. Optional. \cr
}
}

