% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_letter.R
\name{cover_letter}
\alias{cover_letter}
\alias{templ_cover_letter}
\title{Steve's (Non-Academic) Cover Letter Template}
\usage{
cover_letter(...)

templ_cover_letter()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for cover letters when applying for non-academic jobs. For more information, see here:
\url{https://twitter.com/stevenvmiller/status/1320831949469065218}. This template will force the use of
\code{xelatex}.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{author} \tab name of the author \cr
   \code{address} \tab the author's address, which can (and should) be piped with \code{|} separators \cr
   \code{email} \tab the author's email \cr
   \code{url} \tab the author's website domain \cr
   \code{phone} \tab the author's phone number \cr
   \code{fontawesome} \tab logical, defaults to \code{TRUE}. If \code{TRUE}, use \code{fontawesome} icons \cr
   \code{mainfont} \tab The main font to use with \code{xelatex}. I use \code{cochineal} here. Optional, but recommended. \cr
   \code{monofont} \tab The monospaced font to use with \code{xelatex}. I use \code{Fira Code} here. Optional, but recommended. \cr
   \code{date} \tab Specify a date format \cr
   \code{greetings} \tab Specify a salutations to start the letter \cr
}
}

