% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\alias{templ_cv}
\title{Steve's Academic CV Template}
\usage{
cv(...)

templ_cv()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for academic CVs. For more information, see here:
\url{http://svmiller.com/blog/2016/03/svm-r-markdown-cv/}.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{author} \tab name of the author \cr
   \code{jobtitle} \tab appears as first row in the CV \cr
   \code{address} \tab appears as second row in the CV \cr
   \code{fontawesome} \tab logical, defaults to \code{TRUE}. If \code{TRUE}, use fontawesome icons \cr
   \code{email} \tab your email, for the third row \cr
   \code{github} \tab optional, displays Github user name on third row \cr
   \code{orcid} \tab optional, displays ORCID ID on third row \cr
   \code{phone} \tab optional, displays your phone number on third row \cr
   \code{web} \tab optional, displays your domain name on third row \cr
   \code{updated} \tab optional, displays (on third row) when file was last updated \cr
   \code{keywords} \tab not terribly useful, but some keywords to embed in PDF so Google may find it \cr
}
}

