% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekohist_thesis.R
\name{ekohist_thesis}
\alias{ekohist_thesis}
\title{Stockholm University EKOHIST Thesis Template}
\usage{
ekohist_thesis(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
This is a specialty LaTeX R Markdown template for which I am not anticipating
there will be high demand. This is for master's students in the Department of
Economic History and International Relations at Stockholm University, which
currently employs me and is always welcome to give me more money.
}
\details{
If your version of TeX comes by way of Yihui Xie's \pkg{tinytex}, you may
want to additionally install the "xurl" library. Otherwise, you'll get a
vague error about a missing LaTeX dependency.
\code{tinytex::tlmgr_install("xurl")} should suffice. I thank Ian Adams and
Cornelius Hennch for finding this issue and pointing to this solution.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template. There is not much here that is particularly important.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{author} \tab name of the author (i.e. the student) \cr
   \code{supervisor} \tab the name of the supervising professor (i.e. probably me) \cr
   \code{field} \tab the field of study (i.e. typically economic history of IR) \cr
   \code{mainfont} \tab optional(?), but it's the main font to use with \code{xelatex}. \cr
   \code{abstract} \tab include an abstract of the paper here \cr
   \code{keywords} \tab some keywords for the manuscript \cr
   \code{pandocparas} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, keeps Pandoc's \cr
   default paragraph format \tab  \cr
}
}

