% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_article.R
\name{html_article}
\alias{html_article}
\alias{templ_html_article}
\title{Steve's HTML Article Template}
\usage{
html_article(...)

templ_html_article()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:word_document]{rmarkdown::word_document}}.}
}
\description{
This is the R Markdown template I added in support of another package, which facilitates
the conversion of academic articles to HTML format.
}
