% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memo}
\alias{memo}
\title{Steve's Memo Template}
\usage{
memo(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for memos. For more information, see here:
\url{http://svmiller.com/blog/2019/06/r-markdown-memo-template/}.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{from} \tab name of the author \cr
   \code{to} \tab the intended recipient \cr
   \code{subject} \tab the title of the memo \cr
   \code{date} \tab the date of the memo \cr
   \code{memorandum} \tab logical. If \code{TRUE}, includes MEMORANDUM on top of the memo \cr
   \code{graphics} \tab logical. If \code{TRUE}, allows for graphic logo on memo \cr
   \code{width} \tab adjustable with for logo, if \code{graphics: TRUE} \cr
   \code{logoposition} \tab position of logo, if \code{graphics: TRUE} \cr
   \code{logo} \tab file for logo, if \code{graphics: TRUE} \cr
}
}

