% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{resume}
\alias{resume}
\title{Steve's Non-Academic Résumé Template:}
\usage{
resume(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for non-academic résumés . For more information, see here:
\url{http://svmiller.com/blog/2020/09/rmarkdown-template-non-academic-resume/}.
Do note that \code{xelatex} is functionally required for this template.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{author} \tab name of the author \cr
   \code{fontawesome} \tab logical, defaults to \code{TRUE}. If \code{TRUE}, use fontawesome icons \cr
   \code{email} \tab your email \cr
   \code{location} \tab your address \cr
   \code{phone} \tab your phone number \cr
   \code{github} \tab your Github username (optional) \cr
   \verb{web:} \tab your domain name \cr
   \code{twitter} \tab your Twitter handle \cr
   \code{linkedin} \tab your LinkedIn ID \cr
   \code{includephoto} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, includes photo of you in top-right of document. \cr
   \code{myphoto} \tab file name for your photo, only necessary if \code{includephoto: TRUE}. \cr
   \code{shift} \tab shift/adjust photo, defaults to '7in, -.25in' if \code{includephoto: TRUE} \cr
   \code{photobigness} \tab how big photo should be, adjust to taste if \code{includephoto: TRUE} \cr
   \code{photozoom} \tab adjust zoom on photo, defaults to '.2\textwdith' if \code{includephoto: TRUE} \cr
}
}

