% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{statement}
\alias{statement}
\alias{templ_statement}
\title{Steve's Statement Template}
\usage{
statement(...)

templ_statement()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for what I term "statements." These are things like research statements and teaching statements when I'm
applying for jobs or doing end-of-the-semester busy work for the proverbial "ball and chain."
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab title of the document, appears on first row \cr
   \code{author} \tab name of the author, appears on second row \cr
   \code{affiliation} \tab institutional affiliation, appears on third row \cr
   \code{email} \tab your email, for the fourth row \cr
}
}

