% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus2.R
\name{syllabus2}
\alias{syllabus2}
\alias{templ_syllabus2}
\title{Steve's 2nd Syllabus Template}
\usage{
syllabus2(...)

templ_syllabus2()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
This is the R Markdown template I use for my syllabi at Stockholm University.
Classes are structured a lot differently here, so a lot of what I'd like to
do or aspire to do has to change.
}
\section{About YAML header fields}{
\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab Title of the class \cr
   \code{author} \tab Name(s) of the professor(s) \cr
   \code{term} \tab The term of the class (e.g. Fall 2024, Spring 2025, Summer 2025) \cr
   \code{dates} \tab A basic date range for the course (e.g. Feb. 20 - March 21) \cr
   \code{fontawesome} \tab Logical, defaults to TRUE, for use of Font Awesome icons \cr
   \code{email} \tab The email for the "course responsible" professor. It's a thing here. \cr
   \code{web} \tab Class website. Technically optional, but you really should use it. \cr
   \code{office} \tab Your office location. Technically optional, but you really should use it. \cr
   \code{one-teacher} \tab Specify this argument in the YAML as TRUE for classes with one professor. \cr
   \code{two-teachers} \tab Specify this argument in the YAML as TRUE for classes with two professors. Has implications for formatting. \cr
   \code{n-teachers} \tab Specify this argument in the YAML as TRUE for classes with more than two professors. Has implications for formatting. \cr
}
}

\section{Additional Comments}{
XeLaTeX is imposed by a standard use in RStudio. The one-/two-/n-teacher(s) hack
is inspired by hackery in \code{article3} in this same package. Specify only one
of those.

The Font Awesome package in LaTeX is \code{fontawesome5}.

The "author" argument should be used as if there were potentially multiple
professors (even if there is just the one). You'll see what I mean in the
skeleton file.
}

