% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word}
\alias{word}
\alias{templ_word}
\title{Steve's Word Manuscript Template}
\usage{
word(...)

templ_word()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:word_document]{rmarkdown::word_document}}.}
}
\description{
This is the R Markdown template I use for journals that insist on Microsoft Word submissions.
The functionality is limited, mostly because Microsoft Word is limited. The only real use
here is the effective anonymity that emerges in the Word document when the primary
workflow is still built around compilation to a LaTeX PDF.
}
