% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article2.R
\name{article2}
\alias{article2}
\alias{templ_article2}
\title{Steve's 2nd Article/Manuscript Template}
\usage{
article2(...)

templ_article2()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
Another template for academic articles/manuscripts. For more information, see here:
\url{https://svmiller.com/blog/2020/09/another-rmarkdown-article-template/}. Please note
that \code{xelatex} is functionally required for this template/function.
}
\details{
If your version of TeX comes by way of Yihui Xie's \pkg{tinytex}, you'll want to
additionally install the "xurl" library. Otherwise, you'll get a vague error about a
missing LaTeX dependency. \code{tinytex::tlmgr_install("xurl")} should suffice. I
thank Ian Adams and Cornelius Hennch for finding this issue and pointing to this
solution.

Going forward, I recommend the Quarto version of this template for your projects.
You can find that template here, named \code{qarticle2}:

\url{https://github.com/svmiller/quarto-templates/}
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template. Much should've been carried over from my
first article/manuscript template. Please yell at me/raise an issue
on the Github repo if I didn't.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{name} \tab name of the author \cr
   \code{affiliation} \tab the affiliation of the author \cr
   \code{thanks} \tab A place to thank people for their input, identify corresponding author, etc. \cr
   \code{abstract} \tab include an abstract of the paper here \cr
   \code{keywords} \tab some keywords for the manuscript \cr
   \code{anonymous} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, scrubs all identifying author info for peer review \cr
   \code{removetitleabstract} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, removes title/abstract for peer review (if you'd like) \cr
   \code{appendix} \tab optional, but specifies appendix figure/table prefices \cr
   \code{endnotes} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, makes R Markdown footnotes as endnotes \cr
   \code{pandocparas} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, keeps Pandoc's default paragraph format \cr
   \code{acmsans} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, better mimics ACM's use of sans fonts for section headers \cr
   \code{mainfont} \tab optional (I think?), but it's the main font to use with \code{xelatex}. I use \code{cochineal} here. \cr
   \code{sansfont} \tab optional, only necessary if \code{sansitup: TRUE}. Specifies sans font. I recommend Linux Biolinum or Libertinus Sans \cr
   \code{doublespacing} \tab optional, if anything is here to process, this will double-space the document. \cr
   \code{endfloat} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, figures/tables moved to back of document with "Figure/Table about here" placeholder text. Observed only when \code{anonymous: TRUE} and depends on figure/table captions. \cr
}
}

