% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article2.R
\name{article2}
\alias{article2}
\title{Steve's 2nd Article/Manuscript Template}
\usage{
article2(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
Another template for academic articles/manuscripts. For more information, see here:
\url{http://svmiller.com/blog/2020/09/another-rmarkdown-article-template/}. Please note
that \code{xelatex} is functionally required for this template/function.
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template. Much should've been carried over from my
first article/manuscript template. Please yell at me/raise an issue
on the Github repo if I didn't.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{name} \tab name of the author \cr
   \code{affiliation} \tab the affiliation of the author \cr
   \code{thanks} \tab A place to thank people for their input, identify corresponding author, etc. \cr
   \code{abstract} \tab include an abstract of the paper here \cr
   \code{keywords} \tab some keywords for the manuscript \cr
   \code{anonymous} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, scrubs all identifying author info for peer review \cr
   \code{removetitleabstract} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, removes title/abstract for peer review (if you'd like) \cr
   \code{appendix} \tab optional, but specifies appendix figure/table prefices \cr
   \code{endnotes} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, makes R Markdown footnotes as endnotes \cr
   \code{pandocparas} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, keeps Pandoc's default paragraph format \cr
   \code{sansitup} \tab logical, defaults to \code{FALSE}. If \code{TRUE}, better mimics ACM's use of sans fonts for section headers \cr
   \code{mainfont} \tab optional (I think?), but it's the main font to use with \code{xelatex}. I use \code{cochineal} here. \cr
   \code{sansfont} \tab optional, only necessary if \code{sansitup: TRUE}. Specifies sans font. I recommend Linux Biolinum \cr
}
}

