% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{beamer}
\alias{beamer}
\alias{templ_beamer}
\title{Steve's Beamer Template}
\usage{
beamer(...)

templ_beamer()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:beamer_presentation]{rmarkdown::beamer_presentation}}.}
}
\description{
This is the R Markdown template I use for in-class or in-person presentations. For
a discussion of some of its features, see here:
\url{http://svmiller.com/blog/2019/08/r-markdown-template-beamer-presentations/}. Note
that \code{xelatex} is not required for this template but is heavily suggested.
}
\section{About YAML header fields}{
\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab title of the presentation \cr
   \code{subtitle} \tab subtitle of the presentation. Optional, but useful. For in-class lectures, I put the class name here. \cr
   \code{institute} \tab Optional, but list your department/college here. \cr
   \code{titlegraphic} \tab Optional, but put a university watermark here. \cr
   \code{scaletg} \tab Optional, but scale \code{titlegraphic} relative to the width of the document. Defaults to .3 \cr
   \code{make149} \tab Optional, but gives 14:9 dimension to presentation (my favorite) \cr
   \code{mainfont} \tab Optional but it's the main font to use with \code{xelatex}. I use \verb{Open Sans} here. \cr
   \code{titlefont} \tab Optional for fancy fonts for slide titles. I use \verb{Titillium Web} here. \cr
   \code{primarycolor} \tab Optional, but specifies a "primary" color in six-digit RGB color. Defaults to \verb{522D80} \cr
   \code{secondarycolor} \tab Optional, but specifies a "secondary" color in six-digit RGB color. Defaults to \code{F56600} \cr
}
}

