% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus.R
\name{syllabus}
\alias{syllabus}
\alias{templ_syllabus}
\title{Steve's Syllabus Template}
\usage{
syllabus(...)

templ_syllabus()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
This is the R Markdown template I use for my syllabi. For
a discussion of some of its features, see here:
\url{http://svmiller.com/blog/2016/07/r-markdown-syllabus/}. The skeleton also
includes code for a calendar. I discuss this here:
\url{http://svmiller.com/blog/2020/08/a-ggplot-calendar-for-your-semester/}
}
\section{About YAML header fields}{
\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab Title of the class \cr
   \code{author} \tab Name(s) of the professor(s) \cr
   \code{date} \tab The semester \cr
   \code{email} \tab Your email. Technically optional, but you really should use it. \cr
   \code{web} \tab Class website. Technically optional, but you really should use it. \cr
   \code{officehours} \tab Your office hours. Technically optional, but you really should use it. \cr
   \code{office} \tab Your office location. Technically optional, but you really should use it. \cr
   \code{classroom} \tab The classroom location. Technically optional, but you really should use it. \cr
   \code{classhours} \tab When/for how long the class meets in a given session. Technically optional, but you really should use it. \cr
   \code{coteaching} \tab If TRUE, changes format to allow for a second (third, or fourth) co-teacher \cr
}
}

\section{Additional Comments}{
There is a nested conditionality for co-teachers. If \code{coteaching} is TRUE,
format changes to allow a second teacher. Conditional on \code{coteaching} being
TRUE \emph{and} there being an \code{email3} entry, a third co-teacher is added and
entries for the third teacher's office hours and office are assumed. If
\code{coteaching} is TRUE, \code{email3} has an entry \emph{and} there is an \code{email4} entry,
a fourth co-teacher is added.
}

